CRIMRE-416 Implement warden hook pattern
