+++ app/controllers/application_controller.rb
@@ -9,11 +9,6 @@ class ApplicationController < ActionController::Base
 
   # Redirect user managers to manage users' admin on sign in
   def after_sign_in_path_for(user)
-    Authorising::Revive.new(
-      user: user,
-      user_manager_id: current_user_id
-    ).call
-
     return admin_manage_users_root_path if user.can_manage_others?
 
     super
